removed changes for form completion and sendAndWait from waiting webhooks
